requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.8
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
googleapis-common-protos==1.59.1
gpt4all==0.3.4
greenlet==2.0.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
langchain==0.0.203
langchainplus-sdk==0.0.10
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
oauthlib==3.2.2
openapi-schema-pydantic==1.2.4
packaging==23.1
passlib==1.7.4
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
PyJWT==2.7.0
pyparsing==3.0.9
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
yarl==1.9.2
youtube-transcript-api==0.6.1
